requirements.txt
fastapi
uvicorn[standard]
python-dotenv
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface 